initital commit
this code uses velocity as the panning email item passes transition
areas to determine which action to take (e.g. background should go from
gray to yellow, or yellow back to gray). But the pan gesture can
actually skip points along a path, so it is not possible to use this
method. I need to refactor and just check if translation is within a
range.