all required milestones now functioning as planned, color transitions have been animated for smoothness
There are some if statements inconsistently applied that are intended
to improve efficiency (e.g. only change the background to brown if it
is not already brown). Tim said that the system executes the code, but
is also smart enough not to do anything when it knows nothing is
changed, so I’m unsure if my statements help or not.